apps/ai/requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.8.0
